requirements.txt
AfricastalkingGateway==1.7
appdirs==1.4.3
chardet==2.3.0
diff-match-patch==20121119
Django==1.10.6
django-chartit==0.2.9
django-ckeditor==5.2.1
django-csvimport==2.5
django-import-export==0.5.1
django-jet==1.0.5
django-mail-templated==2.6.5
django-registration==2.2
django-render-block==0.5
django-tastypie==0.13.3
django-templated-email==2.1
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
requests==2.13.0
six==1.10.0
tablib==0.11.4

